requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
boto3==1.9.253
botocore==1.12.253
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
docutils==0.14
exrex==0.10.5
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
joblib==0.14.0
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.1.1
more-itertools==7.2.0
numpy==1.16.5
packaging==19.2
pandas==0.24.2
pandas-datareader==0.8.1
patsy==0.5.1
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.2.2
python-dateutil==2.8.0
pytz==2019.3
regime-switch-model==0.1.1
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.2.2
seaborn==0.9.0
six==1.12.0
soupsieve==1.9.5
statsmodels==0.10.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.6.0
